Добавил данных эксперимента

+++ algorithms/deep_q_networks.py
@@ -47,7 +47,7 @@ class DQN(nn.Module):
 
     def get_padded_to_action_n(self, guide: torch.Tensor):
         if len(guide) == 0:
-            return torch.zeros(self.q_function.action_n)
+            return torch.zeros(self.q_function.action_n, dtype=torch.long)
         return torch.nn.functional.pad(guide,
                                        pad=(0, self.q_function.action_n - guide.numel()),
                                        mode="constant",
